pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["NOOR OMARI <nooromari847@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.1"
psycopg2-binary = "^2.9.1"
gunicorn = "^20.1.0"
whitenoise = "^5.2.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
flake8 = {version = "^3.9.2", allow-prereleases = true}
black = {version = "^21.6b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
